Make training pipeline cleanly runnable end-to-end (fixes #12)
- Use .SECONDARY instead of .INTERMEDIATE so checkpoints survive
- Extract precompute_features.py as separate Makefile step with sentinel
- Pin Lightning checkpoint path (enable_version_counter=False, default_root_dir)
- Simplify bundle_structured_classifier to use fixed checkpoint path
- Remove precompute logic from train_scoring_layer (uses precomputed memmaps)
- Filter empty-records examples in train_structured_classifier
- Update false positive rate threshold for non-union test examples
- Retrained weights with consolidated training data
