--- training/data/temperatures.json
@@ -1,9 +1,0 @@
-{
-  "union_name": 1.2668849229812622,
-  "desig_name": 1.2697746753692627,
-  "desig_num": 2.0268642902374268,
-  "prefix": 1.8128230571746826,
-  "suffix": 1.9345782995224,
-  "f_num": 0.8896189332008362,
-  "scoring": 0.6389
-}
